+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * feat(projects): add edit project event
 * feat(projects): add delete project event
 * feat(projects): add new project created event
+* feat(projects): add update app project event
 
 ## 8.1.0